Create explicit tester for no diagnostics
Created a new analyzer tester to explicitly test for no diagnostics in the code under test.

The tester was based on the original implementation of the analyzer tester: `CSharpAnalyzerTest`.

The main (and only) difference is that this tester ensures that absolutely no diagnostics are reported on the code under test.

The reason why this is required is that the original tester has no way of explicitly being told that no diagnostics should be reported on the code under test.

The original tester's success criteria relied solely on diagnostic matching (matching the reported diagnostics to the diagnostics passed to the tester or the diagnostics marked in the code under test), potentially overlooking instances where the code under test should yield no diagnostics.