Add tester to check for all NRT diagnostics
Creates an analyzer tester that verifies diagnostics at interaction points between nullable-enabled and nullable-disabled code.

This analyzer tester is different from the default tester because it uses a preprocessor directive to conditionally compile the code into two compilations: the actual 'code under test' and the simulated 'equivalent code'. The tester uses the diagnostics reported by the 'equivalent code' to determine what diagnostics should be reported by the actual 'code under test'.

Note that a lot of tests started failing because of this new analyzer tester. These will be fixed in future changes to the analyzer.