Use location mapping to get diagnostic locations
Switches the algorithm to get the diagnostics to report to the user from using the annotated nodes to simply getting all the reported diagnostics in the modified compilation and mapping the diagnostic locations to the original syntax tree.

The location mapping was done by getting all the text changes between the original syntax tree and the modified syntax tree. Each text change is applied in reverse on the location of the diagnostic to get the location in the original syntax tree.